streamhttp/client/goog-adk-client/sqlite-agent/pyproject.toml
[project]
name = "sqlite-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.8"
dependencies = [
    "google-adk>=1.5.0",
    "google-generativeai>=0.8.5",
]
